[refactor] organize Model\Join (#680)
